requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
flask-marshmallow==0.15.0
marshmallow==3.20.1
werkzeug>=3.0.1
python-dotenv==1.0.1
azure-monitor-opentelemetry==1.6.2
azure-monitor-opentelemetry-exporter==1.0.0b31
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-logging==0.48b0
healthcheck==1.3.3